# Speed up record inclusion check.