--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -24,8 +24,8 @@
 #### :bug: Bug Fix
 
 - Fix issue where uncurried type internals leak in type error. https://github.com/rescript-lang/rescript-compiler/pull/6264
-- Improve error messages for untagged variant definitions https://github.com/rescript-lang/rescript-compiler/pull/6290
-- Fix type checking performance issue for large records. https://github.com/rescript-lang/rescript-compiler/issues/6284
+- Improve error messages for untagged variant definition. https://github.com/rescript-lang/rescript-compiler/pull/6290
+- Fix type checking performance issue for large records. https://github.com/rescript-lang/rescript-compiler/pull/6289
 
 # 11.0.0-beta.1