--- a/jscomp/build_tests/super_errors/fixtures/RecordInclusion.res
+++ b/jscomp/build_tests/super_errors/fixtures/RecordInclusion.res
@@ -1,0 +1,5 @@
+module M : {
+  type t<'a, 'b, 'c> = {x:int, y:list<('a, 'b)>, z:int}
+} = {
+  type t<'a, 'b, 'c> = {x:int, y:list<('a, 'c)>, z:int}
+}