Speed up record inclusion check.
Fixes https://github.com/rescript-lang/rescript-compiler/issues/6284

Record inclusion check (between implementation and interface) is quadratic.

Example:
```res
module M : {
  type t<'a, 'b, 'c> = {x:int, y:list<('a, 'b)>, z:int}
} = {
  type t<'a, 'b, 'c> = {x:int, y:list<('a, 'c)>, z:int}
}
```

The existing algorithm tries to instantiate type parameters. It only reports an error if there is an inconsistency. This requires solving type equations involving many types at once.

To give an accurate error message, the first problematic field is reported (in this case `y`). So the type equations are checked again and again with size 1, 2, ...n where n is the number of fields. (Plus the type parameters).

This is quadratic and is problematic for records of ~1K fields.

This PR provides a fast path which just checks if there is an error, without blaming a specific field. The fast path is linear.
Only if an error is detected, the quadratic path is take to blame precisely which field is involved.