@cknitt 2023-06-06
Great work! 🎉 

There is another place where compare_records is invoked (in compare_constructor_arguments). Should that be adapted, too?