@zth 2023-06-06
@mununki could you check what https://github.com/mununki/benchmark-rescript-records gives when running this PR?