@mununki 2023-06-07
> @mununki could you check what [mununki/benchmark-rescript-records](https://github.com/mununki/benchmark-rescript-records) gives when running this PR?

It's clear that performance has improved under the same conditions.
```
# before
Normal record x 7.29 ops/sec ±0.96% (23 runs sampled)
Record with optional x 5.08 ops/sec ±3.61% (17 runs sampled)
Record with spread x 5.30 ops/sec ±1.20% (18 runs sampled)
Record with spread and optional x 3.37 ops/sec ±0.96% (13 runs sampled)

# after
Normal record x 8.55 ops/sec ±0.83% (26 runs sampled)
Record with spread x 6.57 ops/sec ±2.79% (21 runs sampled)
Record with optional x 7.97 ops/sec ±1.18% (25 runs sampled)
Record with spread and optional x 6.40 ops/sec ±0.46% (20 runs sampled)
```