@cristianoc 2023-06-07
> Great work! 🎉
> 
> There is another place where compare_records is invoked (in compare_constructor_arguments). Should that be adapted, too?

Refactored so there's a single function now.