@cristianoc 2023-06-07
One alternative would be to make this internal function incremental:
```ml
let eqtype_list rename type_pairs subst env tl1 tl2 =
  univar_pairs := [];
  let snap = Btype.snapshot () in
  try eqtype_list rename type_pairs subst env tl1 tl2; backtrack snap
  with exn -> backtrack snap; raise exn
```
so it would not slow down even for the error case. However it would be more complicated.